backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.1.0
email-validator==2.1.0

# LangChain & AI
langchain==0.1.0
langchain-community==0.0.20
dashscope==1.14.1
chromadb==0.4.18

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
unstructured==0.11.2
markdown==3.5.1

# Background Tasks
apscheduler==3.10.4

# Rate Limiting
slowapi==0.1.9

# Monitoring & Logging
prometheus-client==0.19.0

# Utilities
tenacity==8.2.3
httpx==0.25.2
python-dateutil==2.8.2
